+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/response/ChecklistWithScoreReadResponse.java
@@ -1,7 +1,6 @@
 package com.bang_ggood.checklist.dto.response;
 
 import com.bang_ggood.checklist.domain.Checklist;
-import com.bang_ggood.checklist.dto.CategoryScoreReadResponse;
 import java.util.List;
 
 public record ChecklistWithScoreReadResponse(
