+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/response/ChecklistsWithScoreReadResponse.java
@@ -1,4 +1,4 @@
-package com.bang_ggood.checklist.dto;
+package com.bang_ggood.checklist.dto.response;
 
 import com.bang_ggood.checklist.dto.response.ChecklistWithScoreReadResponse;
 import java.util.List;
