+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/response/WrittenChecklistResponse.java
@@ -1,4 +1,4 @@
-package com.bang_ggood.checklist.dto;
+package com.bang_ggood.checklist.dto.response;
 
 import com.bang_ggood.category.dto.WrittenCategoryQuestionsResponse;
 import com.bang_ggood.room.dto.WrittenRoomResponse;
