+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/response/WrittenQuestionResponse.java
@@ -1,4 +1,4 @@
-package com.bang_ggood.checklist.dto;
+package com.bang_ggood.checklist.dto.response;
 
 public record WrittenQuestionResponse(Integer questionId, String title, String subtitle, String answer) {
 }
