+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/repository/ChecklistRepository.java
@@ -4,12 +4,10 @@ import com.bang_ggood.checklist.domain.Checklist;
 import com.bang_ggood.exception.BangggoodException;
 import com.bang_ggood.exception.ExceptionCode;
 import com.bang_ggood.user.domain.User;
-
-import java.util.Optional;
-import java.util.List;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import java.util.List;
+import java.util.Optional;
 
 public interface ChecklistRepository extends JpaRepository<Checklist, Long> {
 
@@ -23,7 +21,7 @@ public interface ChecklistRepository extends JpaRepository<Checklist, Long> {
     default Checklist getById(long id) {
         return findById(id).orElseThrow(() -> new BangggoodException(ExceptionCode.USER_NOT_FOUND));
     }
-  
+
     List<Checklist> findByUser(User user);
   
     List<Checklist> findByUserAndIdIn(User user, List<Long> checklistIds);
